requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
jinja2>=3.1.4 
langchain-huggingface>=0.2.0
langchain-openai>=0.2.0
langfuse>=0.1.0